Use bigInteger column type for user_id columns

+++ database/migrations/2016_06_01_000002_create_oauth_access_tokens_table.php
@@ -15,7 +15,7 @@ class CreateOauthAccessTokensTable extends Migration
     {
         Schema::create('oauth_access_tokens', function (Blueprint $table) {
             $table->string('id', 100)->primary();
-            $table->integer('user_id')->index()->nullable();
+            $table->bigInteger('user_id')->index()->nullable();
             $table->unsignedInteger('client_id');
             $table->string('name')->nullable();
             $table->text('scopes')->nullable();
